+++ src/model/easing/Cubic.java
@@ -2,15 +2,15 @@ package model.easing;
 
 public class Cubic extends Easing {
 	
-	public  float easeIn (float t,float b , float c, float d) {
+	public  double easeIn (double t,double b , double c, double d) {
 		return c*(t/=d)*t*t + b;
 	}
 	
-	public  float easeOut (float t,float b , float c, float d) {
+	public  double easeOut (double t,double b , double c, double d) {
 		return c*((t=t/d-1)*t*t + 1) + b;
 	}
 	
-	public  float easeInOut (float t,float b , float c, float d) {
+	public  double easeInOut (double t,double b , double c, double d) {
 		if ((t/=d/2) < 1) return c/2*t*t*t + b;
 		return c/2*((t-=2)*t*t + 2) + b;
 	}
