+++ src/model/easing/Linear.java
@@ -2,15 +2,15 @@ package model.easing;
 
 public class Linear extends Easing {
 	
-	public  float easeIn (float t,float b , float c, float d) {
+	public  double easeIn (double t,double b , double c, double d) {
 		return c*t/d + b;
 	}
 	
-	public  float easeOut (float t,float b , float c, float d) {
+	public  double easeOut (double t,double b , double c, double d) {
 		return c*t/d + b;
 	}
 	
-	public  float easeInOut (float t,float b , float c, float d) {
+	public  double easeInOut (double t,double b , double c, double d) {
 		return c*t/d + b;
 	}
